Extract transaction test helper functions to a separate file
