+++ packages/buidler-core/test/internal/buidler-evm/provider/modules/eth.ts
@@ -1,5 +1,4 @@
 import { assert } from "chai";
-import { Transaction } from "ethereumjs-tx";
 import { BN, bufferToHex, toBuffer, zeroAddress } from "ethereumjs-util";
 import { Context } from "mocha";
 
@@ -13,9 +12,7 @@ import {
   RpcReceiptOutput,
   RpcTransactionOutput,
 } from "../../../../../src/internal/buidler-evm/provider/output";
-import { BuidlerEVMProvider } from "../../../../../src/internal/buidler-evm/provider/provider";
 import { getCurrentTimestamp } from "../../../../../src/internal/buidler-evm/provider/utils/getCurrentTimestamp";
-import { EthereumProvider } from "../../../../../src/types";
 import {
   assertInvalidInputError,
   assertNodeBalances,
@@ -37,82 +34,20 @@ import {
 } from "../../helpers/conversions";
 import { setCWD } from "../../helpers/cwd";
 import {
-  DEFAULT_ACCOUNTS,
   DEFAULT_ACCOUNTS_ADDRESSES,
   DEFAULT_ACCOUNTS_BALANCES,
   DEFAULT_BLOCK_GAS_LIMIT,
   PROVIDERS,
 } from "../../helpers/providers";
-import { retrieveCommon } from "../../helpers/retrieveCommon";
 import { retrieveForkBlockNumber } from "../../helpers/retrieveForkBlockNumber";
+import {
+  deployContract,
+  getSignedTxHash,
+  sendTransactionFromTxParams,
+  sendTxToZeroAddress,
+} from "../../helpers/transactions";
 
 const PRECOMPILES_COUNT = 8;
-
-async function sendTxToZeroAddress(
-  provider: EthereumProvider
-): Promise<string> {
-  const accounts = await provider.send("eth_accounts");
-
-  const burnTxParams = {
-    from: accounts[0],
-    to: zeroAddress(),
-    value: numberToRpcQuantity(1),
-    gas: numberToRpcQuantity(21000),
-    gasPrice: numberToRpcQuantity(1),
-  };
-
-  return provider.send("eth_sendTransaction", [burnTxParams]);
-}
-
-async function deployContract(
-  provider: EthereumProvider,
-  deploymentCode: string
-) {
-  const hash = await provider.send("eth_sendTransaction", [
-    {
-      from: DEFAULT_ACCOUNTS_ADDRESSES[0],
-      data: deploymentCode,
-      gas: numberToRpcQuantity(DEFAULT_BLOCK_GAS_LIMIT),
-    },
-  ]);
-
-  const { contractAddress } = await provider.send("eth_getTransactionReceipt", [
-    hash,
-  ]);
-
-  return contractAddress;
-}
-
-async function sendTransactionFromTxParams(
-  provider: EthereumProvider,
-  txParams: TransactionParams
-) {
-  return provider.send("eth_sendTransaction", [
-    {
-      to: bufferToHex(txParams.to),
-      from: bufferToHex(txParams.from),
-      data: bufferToHex(txParams.data),
-      nonce: numberToRpcQuantity(txParams.nonce),
-      value: numberToRpcQuantity(txParams.value),
-      gas: numberToRpcQuantity(txParams.gasLimit),
-      gasPrice: numberToRpcQuantity(txParams.gasPrice),
-    },
-  ]);
-}
-
-async function getSignedTxHash(
-  buidlerEVMProvider: BuidlerEVMProvider,
-  txParams: TransactionParams,
-  signerAccountIndex: number
-) {
-  const txToSign = new Transaction(txParams, {
-    common: await retrieveCommon(buidlerEVMProvider),
-  });
-
-  txToSign.sign(toBuffer(DEFAULT_ACCOUNTS[signerAccountIndex].privateKey));
-
-  return bufferToHex(txToSign.hash(true));
-}
 
 describe("Eth module", function () {
   PROVIDERS.forEach(({ name, useProvider, isFork, chainId }) => {
